requirements.txt
Flask==3.0.3
PyYAML==6.0.2
WTForms==3.1.2
Flask-WTF==1.2.1